Refactor Server._report_config_data and methods to Tenant
